+++ NEWS.md
@@ -1,3 +1,11 @@
+# CHANGES IN tufte VERSION 0.3 (unreleased)
+
+## NEW FEATURES
+
+- It is possible to select a subset of some features of the default Tufte style (`tufte-css`) via the `tufte_features` argument of `tufte_html()`. For example, you can disable the `et-book` fonts and the background color.
+
+- A new variant of the Tufte style, `envisioned`, is added to `tufte_html()`. You can use `tufte_html(tufte_variant = 'envisioned')` to enable this style. The major difference with the default Tufte style is: the font family is `Roboto Condensed`, the background color is `#fefefe`, and the text color is `#222` (thanks, @eddelbuettel, #21).
+
 # CHANGES IN tufte VERSION 0.2
 
 ## NEW FEATURES